pyproject.toml
[project]
name = "alpamayo_r1"
version = "0.1.0"
requires-python = "==3.12.*"
dependencies = [
  "accelerate>=1.12.0",
  "av>=16.0.1",
  "bitsandbytes>=0.45.0",
  "einops>=0.8.1",
  "hydra-colorlog>=1.2.0",
  "hydra-core>=1.3.2",
  "pandas>=2.3.3",
  "physical_ai_av>=0.1.0",
  "pillow>=12.0.0",
  "torch==2.8.0",
  "torchvision>=0.23.0",
  "transformers==4.57.1",
  "flash-attn>=2.8.3",
]

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
  "matplotlib>=3.10.7",
  "mediapy>=1.2.4",
  "ipykernel>=6.29.3",
  "ipywidgets>=8.1.8",
]

[tool.uv]
no-build-isolation-package = ["flash-attn"]

[tool.ruff]
line-length = 100
